requirements.txt
pre-commit==2.18.1
pytest==7.1.2
mysqlclient==2.1.0
Django==4.0.4
djangorestframework==3.13.1
sqlparse==0.3.1
django-filter==21.1